+++ packages/plugin-packages/mashroom-http-proxy/test/proxy/ProxyImplNodeStreamAPI.pure.test.ts
@@ -370,7 +370,7 @@ describe('ProxyImplNodeStreamAPI', () => {
                         loggerFactory: loggingUtils.dummyLoggerFactory,
                         services: {}
                     } as any;
-                    httpProxy.forwardWs(req, socket, head, 'ws://ws.ifelse.io/', {
+                    httpProxy.forwardWs(req, socket, head, 'wss://ws.ifelse.io/', {
 
                     });
                 });